ai/requirements.txt
fastapi
pydantic
uvicorn
pandas
openai
psycopg2-binary
python-dotenv